ftx/requirements.txt
ciso8601
requests
google-cloud-storage